requirements.txt
# Core dependencies
openai>=1.3.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0

# PDF processing
PyPDF2>=3.0.1
pdfplumber>=0.10.2
PyMuPDF>=1.23.8

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
requests>=2.31.0
python-multipart>=0.0.6
beautifulsoup4>=4.12.2
tiktoken>=0.5.1

# Optional: LangChain for advanced features
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community>=0.0.10

# Development and testing
pytest>=7.4.3
black>=23.11.0
flake8>=6.1.0 